switch to sixe_t ...

+++ pmacct_tcpflags_generic_p2.c
@@ -20,8 +20,8 @@ avro_schema_t sc_type_record, sc_type_array, sc_type_string;
 //
 void compose_tcpflags_avro_schema(void);
 
-unsigned int generate_rnd(void);
-void generate_tcpflags_array(unsigned int);
+size_t generate_rnd(void);
+void generate_tcpflags_array(size_t);
 
 
 int
@@ -31,11 +31,11 @@ main(void)
 
   while(1)
   {
-    unsigned int rnd = generate_rnd();
+    size_t rnd = generate_rnd();
 
     generate_tcpflags_array(rnd);
 
-    unsigned int idx_0;
+    size_t idx_0;
     for (idx_0 = 0; idx_0 < 6; idx_0++)
     {
       printf("%s ", tcpflags[idx_0]);
@@ -48,26 +48,26 @@ main(void)
 }
 
 
-unsigned int
+size_t
 generate_rnd()
 {
-  unsigned int rnd;
+  size_t rnd;
 
-  srand((unsigned int) time(0));
+  srand((size_t) time(0));
   rnd = rand() % 64;
 
-  printf("%u\n", rnd);
+  printf("%lu\n", rnd);
 
   return rnd;
 }
 
 
-void generate_tcpflags_array(unsigned int tcpflags_decimal)
+void generate_tcpflags_array(size_t tcpflags_decimal)
 {
-  unsigned int tcpflags_binary[6] = {0, 0, 0, 0, 0, 0};
+  size_t tcpflags_binary[6] = {0, 0, 0, 0, 0, 0};
   const char tcpflags_mask[6][5] = {"URG", "ACK", "PSH", "RST", "SYN", "FIN"};
 
-  unsigned int idx_0;
+  size_t idx_0;
   if ((tcpflags_decimal > 0) && (tcpflags_decimal) < 64)
   {
     for (idx_0 = 5; tcpflags_decimal > 0 && idx_0 >= 0; idx_0--)
@@ -77,17 +77,17 @@ void generate_tcpflags_array(unsigned int tcpflags_decimal)
     }
   }
 
-  unsigned int idx_1;
+  size_t idx_1;
   for (idx_1 = 0; idx_1 < 6; idx_1++)
   {
     if (!tcpflags_binary[idx_1])
     {
-      printf("%u ", tcpflags_binary[idx_1]);
+      printf("%lu ", tcpflags_binary[idx_1]);
       strcpy(tcpflags[idx_1], "NULL");
     }
     else
     {
-      printf("%u ", tcpflags_binary[idx_1]);
+      printf("%lu ", tcpflags_binary[idx_1]);
       strcpy(tcpflags[idx_1], tcpflags_mask[idx_1]);
     }
   }
